Noncoprime check skipped for unique prime iterators.
Added an inner class PrimeSampler inside the CRA domain that
samples unique primes from a prime generator, checking non-divisibility
with the underlying builder, but skips this when it's a unique
prime generator.
(relates to issue #71)

Also needed to create an adaptor class so that a BlasVector
(or std::vector) of primes can be treated like a PrimeIterator.
This is the PrimeSequence class added to random-prime.h.